move client to pkg, helper script for self-signed certs & update documentation
