+++ pkg/dclient/client_test.go
