+++ pkg/dclient/utils.go
